WilliamsonTransform

Documentation for WilliamsonTransform.

WilliamsonTransform.𝒲

WilliamsonTransform.taylor

WilliamsonTransform.𝒲 — Type

𝒲(X,d)(x)

Computes the Williamson d-transform of the random variable X, taken at point x, as described in

WILLIAMSON, R. E. (1956). Multiply monotone functions and their Laplace transforms. Duke Math. J. 23 189–207. MR0077581

and used in

McNeil, Alexander J., and Johanna Nešlehová. "Multivariate Archimedean copulas, d-monotone functions and ℓ 1-norm symmetric distributions." (2009): 3059-3097.

For a univariate non-negative random variable $X$ for distribution function $F$ and $d\ge 2$ and integer, the williamson-d-transform of $X$ is the real function supported on $[0,\infty[$ given by:

\[𝒲_{X,d}(x) = \int_{x}^{\infty} \left(1 - \frac{x}{t}\right)^{d-1} dF(t) = \mathbb E\left( (1 - \frac{x}{X})^{d-1}_+\right) \mathbb 1_{x > 0} + \left(1 - F(0)\right)\mathbb 1_{x <0}\]

WilliamsonTransform.taylor — Method

𝒲₋₁(ϕ,d)

Computes the inverse Williamson d-transform of the d-monotone archimedean generator ϕ. This inverse is a CDF, and we return it on the form of a random variable <:Distributions.ContinuousUnivariateDistribution from Distributions.jl. The result can be sampled through Distributions.rand(). See

WILLIAMSON, R. E. (1956). Multiply monotone functions and their Laplace transforms. Duke Math. J. 23 189–207. MR0077581

and moreover

McNeil, Alexander J., and Johanna Nešlehová. "Multivariate Archimedean copulas, d-monotone functions and ℓ 1-norm symmetric distributions." (2009): 3059-3097.

for details.

The cumulative distribution function of this random variable is given by:

\[𝒲₋₁(X,d)(x) = 1 - \frac{(-x)^{d-1} \phi_+^{(d-1)}(x)}{k!} - \sum_{k=0}^{d-2} \frac{(-x)^k \phi^{(k)}(x)}{k!}\]